requirements.txt
streamlit
pillow
pandas
rembg
opencv-python